Reset switched to shutdown

+++ Runtime/OpenXR/Network/NetworkRuntime.cs
@@ -15,11 +15,11 @@ namespace Twinny.System.Network
         
         [Info("Seconds")]
         [Tooltip("Wait X seconds for connection until start single player experience.")]
-        public int connectionTimeout = 10;
+        public float connectionTimeout = 10;
         public bool tryReconnect = true;
         [Info("Seconds")]
         [Tooltip("When recenter and application was disconnected, wait X seconds until to restart experience.")]
-        public int resetExperienceDelay = 0;
+        public float resetExperienceDelay = 0;
 
     }
 }
